Branch for abstracting the pubsub API in res_pjsip_pubsub.
This is the initial require step in order to be able to implement
Resource List Subscriptions. The main goal here is to make subscription
handlers no longer rely on PJSIP-specific constructs since when RLS is
in use, there may not be a PJSIP subscription directly beneath the
virtual subscription being worked on by the subscription handler.