Commit some progress towards this effort.
Unfortunately, I've simply run out of time for the day
so this is a big garbled mess at the moment. This does
not even come close to compiling right now.

I've made the necessary changes to res_pjsip_pubsub.h, and
I'm working on the pubsub core before I start adjusting
subscription handlers' calls to the API.