accountcode propagation: Change accountcode propagation some more.
* Changed accountcode propagation over local channels (core unreal
channels) to cross over the local bridge just like a normal bridge.

* Fixed a few places dealing with accountcodes that were reading from
channels without the lock held.

Found that the requestor channel to ast_request() is not necessarily a
bridge peer channel.  It could be the channel being replaced by the new
channel.  I also found that the accountcodes were not being propagated
before these accouncode changes for FollowMe and Queue when local channels
were involved.