main/frame: Only unref formats on non-static frames
The format on the frame was being unreffed on static frames that had
not bumped the ref resulting in a crash when compiled with the DO_CRASH
flag due to the refcount on the format going negative.

This also fixes some frame leaks on startup resulting from this patch.

ASTERISK-23960
Review: https://reviewboard.asterisk.org/r/3738/