media formats: Fix format cache case sensitivity
This changes the format cache to be case insensitive which allows the
tests/channels/pjsip/video_calls/fast_picture_update integration test
to pass. It was failing in the media formats branch due to the "alice"
endpoint failing to load because codecs H261 and H263 were not
recognized.

Review: https://reviewboard.asterisk.org/r/3778/