ARI: Add missing transfer information
This adds a new field to the ast_attended_transfer_type() and a new
subtype AST_ATTENDED_TRANSFER_DEST_LOCAL_APP to describe attended
transfers where a local channel is used to connect a dialplan
application to a bridge in cases where a single remote party can not be
moved directly into the application. Previously, the local channel half
being pushed into the bridge in place of a transferer leg was not
conveyed in the message.

ASTERISK-23941
Review: https://reviewboard.asterisk.org/r/3728/