Fix situation where we were writing to freed memory.
Sending a NOTIFY that terminates a subscription results in the
subscription tree's refcount being decremented. Therefore, in
a situation when sending a NOTIFY that may terminate a subscription,
bumping the refcount is necessary prior to sending the NOTIFY.

This was already actually being done, but its scope of the reference
bump was too narrow.